Desktop/Kangacook project/requirements.txt
absl-py==1.4.0
anyio==4.0.0
appnope==0.1.3
asgiref==3.4.1
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black @ git+https://github.com/psf/black@f51e53726b39a177355a7917c91c56f390dda7ef
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diskcache==5.6.1
dj-database-url==0.5.0
dj-rest-auth==2.2.1
Django==4.0.8
django-allauth==0.44.0
django-ckeditor==6.5.1
django-cors-headers==3.10.1
django-dirtyfields==1.9.0
django-environ==0.9.0
django-extensions==3.2.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
django-rest-auth==0.9.5
django-rq==2.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.4.2
einops==0.6.1
exceptiongroup==1.0.4
executing==1.2.0
filelock==3.12.2
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
httpcore==0.18.0
httpx==0.25.0
hypothesis==6.84.3
idna==3.3
iniconfig==1.1.1
install==1.3.5
ipython==8.14.0
-e git+https://github.com/inthedepthsofitall/ivy.git@c9084f84702dc724740fd90f3b0d58de65367e42#egg=ivy
ivy-core==1.1.10
jax==0.4.13
jaxlib==0.4.13
jedi==0.18.2
Jinja2==3.1.2
jsonpickle==3.0.1
keras==2.13.1
libclang==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
mypy-extensions==0.4.3
networkx==3.1
numpy==1.24.3
nvidia-ml-py==11.495.46
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddlepaddle==2.5.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.48
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.4.0
pymongo==4.4.1
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
pyvis==0.3.2
redis==4.3.4
requests==2.27.0
requests-oauthlib==1.3.0
rq==1.11.1
rsa==4.9
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.2
stripe==3.2.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==1.1.0
tomli==1.2.3
torch==2.0.1
torch-scatter==2.1.1
traitlets==5.9.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
wcwidth==0.2.6
Werkzeug==2.3.6
whitenoise==5.3.0
wrapt==1.14.1
